getPersonFullName
